Put all menu items section generators in menuItems.js
