+++ js/webviewMenu/menuSections.js
@@ -1,21 +1,16 @@
-const makeLinkMenuItems = require('./linkMenuItems')
-const makeImageMenuItems = require('./imageMenuItems')
-const makeSelectionMenuItems = require('./selectionMenuItems')
-const makeNavigationMenuItems = require('./navigationMenuItems')
-const makeClipboardMenuItems = require("./clipboardMenuItems")
+const menuItems = require("./menuItems")
 
 module.exports = function makeMenuSections (clipboard, data, searchEngine) {
-  var currentTabIsPrivate = tabs.get(tabs.getSelected()).private
+  const currentTabIsPrivate = tabs.get(tabs.getSelected()).private
 
   var menuSections = []
 
-  var link = data.linkURL || data.frameURL
-
-  var image = data.srcURL
+  const link = data.linkURL || data.frameURL
+  const image = data.srcURL
 
   /* we don't show the image actions if there are already link actions, because it makes the menu too long and because the image actions typically aren't very useful if the image is a link */
-  menuSections.push(makeLinkMenuItems(link, currentTabIsPrivate) 
-                 || makeImageMenuItems(image, currentTabIsPrivate))
+  menuSections.push(menuItems.link(link, currentTabIsPrivate) 
+                 || menuItems.image(image, currentTabIsPrivate))
 
   if (image) {
     menuSections.push([
@@ -29,12 +24,12 @@ module.exports = function makeMenuSections (clipboard, data, searchEngine) {
   }
 
   const selection = data.selectionText
-  menuSections.push(makeSelectionMenuItems(selection, currentTabIsPrivate, searchEngine))
+  menuSections.push(menuItems.selection(selection, currentTabIsPrivate, searchEngine))
 
-  const clipboardActions = makeClipboardMenuItems(link, image, selection, data, clipboard)
+  const clipboardActions = menuItems.clipboard(link, image, selection, data, clipboard)
   menuSections.push(clipboardActions)
 
-  menuSections.push(makeNavigationMenuItems())
+  menuSections.push(menuItems.navigation())
 
   /* inspect element */
   menuSections.push([
